fix: full codebase review — 24 fixes across 21 files
Memory safety:
- VM PUSH macro bounds check (eval.c)
- op_callf arity guards for RAY_UNARY/RAY_BINARY (eval.c)
- COW leak in ray_vec_append and ray_list_append (vec.c, list.c)
- Serde size overflow guard for huge vectors (serde.c)

Null pointer / crash:
- NULL segment guards in all parted column paths (exec.c, filter.c, expr.c)
- IPC client 256MB allocation cap (ipc.c)

Correctness:
- Replace GNU ({...}) statement expressions with C17 (expr.c)
- dl_add_edb error checking on column fetch (datalog.c)
- sum(I32) returns I64 to prevent truncation (agg.c)
- Type inference returns BOOL for comparison/boolean ops (opt.c)
- OOM cleanup in ray_nfo_create (nfo.c)
- exec_filter_head returns error instead of unfiltered input (filter.c)
- sel_compact returns error instead of silently truncating (filter.c)

Important:
- Assert on ray_env_set failure in builtin registration (eval.c)
- parse_vector STR branch leak fix (parse.c)
- dag_can_stream uses scratch_alloc instead of ray_sys_alloc (exec.c)
- epoll/kqueue UAF guard: re-validate selector before data_fn (epoll.c, kqueue.c)
- LFTJ root enumeration scans all CSRs (lftj.c)
- dl_rule_add_cmp_expr updates n_vars from expression trees (datalog.c)
- Pool overflow assert fixes heap_id before assert (heap.c)

Parted RAY_STR hardening:
- Deep-copy STR gather/flatten/head/tail/filter via ray_str_vec_append
- parted_str_append_elem preserves null semantics
- parted_first_attrs scans first non-NULL segment for SYM width
- parted_seg_esz_ok validates per-segment element size
- Multi-pool STR segments handled correctly (no raw memcpy)

Low/latent:
- slab_count type uint32_t (heap.h)
- RAY_IS_ERR check in sym.c
- compile_list n==0 defensive guard (compile.c)
- HIST_MAX_ENTRIES enforced in-memory (term.c)

+++ src/mem/heap.h
@@ -248,7 +248,7 @@ static inline ray_t* ray_buddy_of(ray_t* v, uint8_t order, uintptr_t pool_base) 
  * Slab cache for small blocks (orders 6-10, i.e., 64B-1024B)
  * -------------------------------------------------------------------------- */
 typedef struct {
-    int64_t  count;
+    uint32_t count;
     ray_t*    stack[RAY_SLAB_CACHE_SIZE];
 } ray_slab_t;
